requirements.txt
channels==2.2.0
channels-redis==2.4.0
daphne==2.3.0
Django==2.2.16
django-cors-headers==2.4.0
django-debug-toolbar==1.11
djangorestframework==3.10.3
djangorestframework-simplejwt==3.3
drf-tracking==1.5.0
fabric==2.4.0
gunicorn==19.9.0
mysqlclient==1.4.1
redis==2.10.6
requests==2.21.0
drf-yasg==1.20.0